Code improvements.
Abstract classes moved to tester_dataclasses.
