+++ package_algorithms/__init__.py
@@ -1,4 +1,3 @@
-from algorithm_tester.abstracts import Algorithm
 from package_algorithms.bb import BranchBound, SortedBranchBound
 from package_algorithms.dp import DynamicProgramming, DynamicProgramming_Weight
 from package_algorithms.basic import BruteForce, Greedy
